pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "matrix-llmagent"
version = "1.0.0"
description = "Agentic LLM chatbot for Matrix with persistent memory and tool use"
authors = [
    {name = "pasky", email = "pasky@ucw.cz"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "ddgs>=0.1.0",
    "markdownify>=0.11.0",
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.0",
    "e2b-code-interpreter<2",
    "openai>=1.40.0",
    "matrix-nio[e2e]>=0.24.0",
    "asyncpg>=0.29.0",
    "pillow>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
matrix-llmagent = "matrix_llmagent.main:main"
matrix-invites = "matrix_llmagent.invite_manager:run"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "I",      # isort
]
ignore = [
    "E501",   # line too long, handled by black
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["SIM"]

[tool.pyright]
include = ["matrix_llmagent", "tests"]
exclude = ["**/node_modules", "**/__pycache__"]
typeCheckingMode = "basic"
pythonVersion = "3.11"
venvPath = "."
venv = ".venv"

[tool.hatch.build.targets.wheel]
packages = ["matrix_llmagent"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers -n auto"
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pyright==1.1.407",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-xdist>=3.8.0",
    "ruff==0.14.6",
]
